+++ packages/electricsql/lib/src/satellite/error.dart
@@ -7,12 +7,14 @@ const _kFatalErrorDescription =
 const _throwErrors = {
   SatelliteErrorCode.internal,
   SatelliteErrorCode.fatalError,
+  SatelliteErrorCode.connectionCancelledByDisconnect,
 };
 
 const _fatalErrors = {
   SatelliteErrorCode.invalidRequest,
   SatelliteErrorCode.unknownSchemaVersion,
   SatelliteErrorCode.authRequired,
+  SatelliteErrorCode.authExpired,
 };
 
 const _outOfSyncErrors = [