+++ packages/electricsql/lib/src/sockets/sockets.dart
@@ -8,9 +8,11 @@ import 'package:web_socket_channel/web_socket_channel.dart';
 
 export 'stub.dart'
     if (dart.library.io) 'io.dart'
-    if (dart.library.html) 'html.dart';
+    if (dart.library.js_interop) 'web.dart';
 
 typedef Data = Uint8List;
+
+const kAuthExpiredCloseEvent = 'JWT-expired';
 
 /// Returns the protocol version string as the server expects. i.e: 'electric.0.6'
 const kProtocolVsn = 'electric.$kElectricProtocolVersion';
@@ -22,7 +24,7 @@ abstract class Socket {
 
   void onMessage(void Function(Data data) cb);
   void onError(void Function(SatelliteException error) cb);
-  void onClose(void Function() cb);
+  void onClose(void Function(SocketCloseReason reason) cb);
 
   void onceConnect(void Function() cb);
   void onceError(void Function(SatelliteException error) cb);
@@ -40,6 +42,13 @@ abstract class SocketFactory {
   Socket create(String protocolVsn);
 }
 
+class CloseEvent {
+  final int code;
+  final String reason;
+
+  CloseEvent({required this.code, required this.reason});
+}
+
 /// Socket implementation that uses web_socket_channel
 /// io and html both derive from the main logic here
 abstract class WebSocketBase implements Socket {
@@ -51,7 +60,7 @@ abstract class WebSocketBase implements Socket {
   List<void Function(SatelliteException error)> _errorCallbacks = [];
   List<void Function(SatelliteException error)> _onceErrorCallbacks = [];
 
-  void Function()? _closeListener;
+  void Function(CloseEvent)? _closeListener;
   void Function(Data data)? _messageListener;
 
   WebSocketBase(this.protocolVsn);
@@ -173,14 +182,23 @@ abstract class WebSocketBase implements Socket {
   }
 
   @override
-  void onClose(void Function() cb) {
+  void onClose(void Function(SocketCloseReason reason) cb) {
     if (_closeListener != null) {
       throw SatelliteException(
         SatelliteErrorCode.internal,
         'socket does not support multiple close listeners',
       );
     }
-    _closeListener = cb;
+
+    // ignore: prefer_function_declarations_over_variables
+    final callback = (CloseEvent ev) {
+      final reason = ev.reason == kAuthExpiredCloseEvent
+          ? SocketCloseReason.authExpired
+          : SocketCloseReason.socketError;
+      cb(reason);
+    };
+
+    _closeListener = callback;
   }
 
   @override
@@ -205,9 +223,23 @@ abstract class WebSocketBase implements Socket {
     }
     _subscriptions = [];
 
-    _channel?.sink.close();
+    final channelToClose = _channel;
 
-    _closeListener?.call();
+    if (channelToClose != null) {
+      final closeFuture = channelToClose.sink.close();
+      if (_closeListener != null) {
+        final listener = _closeListener!;
+        closeFuture.then((_) {
+          listener.call(
+            CloseEvent(
+              code: channelToClose.closeCode!,
+              reason: channelToClose.closeReason!,
+            ),
+          );
+        });
+      }
+    }
+
     _closeListener = null;
 
     _onceConnectCallbacks = [];