+++ packages/electricsql/lib/src/util/types.dart
@@ -42,6 +42,7 @@ class SatelliteException implements Exception {
 }
 
 enum SatelliteErrorCode {
+  connectionCancelledByDisconnect,
   connectionFailedAfterRetry,
   internal,
   timeout,
@@ -59,6 +60,7 @@ enum SatelliteErrorCode {
   authError,
   authFailed,
   authRequired,
+  authExpired,
 
   // server errors
   invalidRequest,
@@ -88,6 +90,18 @@ enum SatelliteErrorCode {
   // shape data errors
   shapeDeliveryError,
   shapeSizeLimitExceeded,
+}
+
+class SocketCloseReason {
+  final SatelliteErrorCode code;
+
+  SocketCloseReason._(this.code);
+
+  static final authExpired =
+      SocketCloseReason._(SatelliteErrorCode.authExpired);
+
+  static final socketError =
+      SocketCloseReason._(SatelliteErrorCode.socketError);
 }
 
 class Replication<TransactionType> {
@@ -325,8 +339,16 @@ class TransactionEvent {
   TransactionEvent(this.transaction, this.ackCb);
 }
 
-enum ConnectivityState {
-  available,
+enum ConnectivityStatus {
   connected,
   disconnected,
+}
+
+class ConnectivityState {
+  final ConnectivityStatus status;
+
+  /// reason for `disconnected` status
+  final SatelliteException? reason;
+
+  const ConnectivityState({required this.status, this.reason});
 }