+++ packages/electricsql/test/satellite/process_test.dart
@@ -46,6 +46,7 @@ MockSatelliteClient get client => context.client;
 String get dbName => context.dbName;
 AuthState get authState => context.authState;
 AuthConfig get authConfig => context.authConfig;
+String get token => context.token;
 
 const parentRecord = <String, Object?>{
   'id': 1,
@@ -61,9 +62,10 @@ const childRecord = <String, Object?>{
 Future<({Future<void> connectionFuture})> startSatellite(
   SatelliteProcess satellite,
   AuthConfig authConfig,
+  String token,
 ) async {
   await satellite.start(authConfig);
-  satellite.setToken(insecureAuthToken({'sub': 'test-user'}));
+  satellite.setToken(token);
   final connectionFuture = satellite.connectWithBackoff();
   return (connectionFuture: connectionFuture);
 }
@@ -100,11 +102,12 @@ void main() {
   });
 
   test('set persistent client id', () async {
-    await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     final clientId1 = satellite.authState!.clientId;
+    await conn.connectionFuture;
     await satellite.stop();
 
-    await startSatellite(satellite, authConfig);
+    await startSatellite(satellite, authConfig, token);
 
     final clientId2 = satellite.authState!.clientId;
 
@@ -114,8 +117,43 @@ void main() {
     await Future<void>.delayed(const Duration(milliseconds: 100));
   });
 
+  test('can use user_id in JWT', () async {
+    // Doesn't throw
+    await startSatellite(
+      satellite,
+      authConfig,
+      insecureAuthToken({'user_id': 'test-userA'}),
+    );
+  });
+
+  test('can use sub in JWT', () async {
+    // Doesn't throw
+    await startSatellite(
+      satellite,
+      authConfig,
+      insecureAuthToken({'sub': 'test-userB'}),
+    );
+  });
+
+  test('require user_id or sub in JWT', () async {
+    expect(
+      () => startSatellite(
+        satellite,
+        authConfig,
+        insecureAuthToken({'custom_user_claim': 'test-userC'}),
+      ),
+      throwsA(
+        isA<ArgumentError>().having(
+          (e) => e.message,
+          'message',
+          'Token does not contain a sub or user_id claim',
+        ),
+      ),
+    );
+  });
+
   test('cannot update user id', () async {
-    await startSatellite(satellite, authConfig);
+    await startSatellite(satellite, authConfig, token);
     expect(
       () => satellite.setToken(insecureAuthToken({'sub': 'test-user2'})),
       throwsA(
@@ -170,6 +208,16 @@ void main() {
     final expectedChange = Change(
       qualifiedTablename: const QualifiedTablename('main', 'parent'),
       rowids: [1, 2],
+      recordChanges: [
+        RecordChange(
+          primaryKey: {'id': 1},
+          type: RecordChangeType.insert,
+        ),
+        RecordChange(
+          primaryKey: {'id': 2},
+          type: RecordChangeType.insert,
+        ),
+      ],
     );
 
     expect(changes, [expectedChange]);
@@ -231,7 +279,7 @@ void main() {
 
     await adapter.run(Statement("INSERT INTO parent(id) VALUES ('1'),('2')"));
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     await Future<void>.delayed(opts.pollingInterval);
@@ -252,7 +300,7 @@ void main() {
     // no txn notified
     expect(notifier.notifications.length, 4);
 
-    final conn1 = await startSatellite(satellite, authConfig);
+    final conn1 = await startSatellite(satellite, authConfig, token);
     await conn1.connectionFuture;
     await Future<void>.delayed(opts.pollingInterval);
 
@@ -1313,16 +1361,16 @@ void main() {
     expect(opLog, expected);
   });
 
-  test('handling connectivity state change stops queueing operations',
-      () async {
+  test('disconnect stops queueing operations', () async {
     await runMigrations();
-    await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
 
     await adapter.run(
       Statement(
         "INSERT INTO parent(id, value, other) VALUES (1, 'local', 1)",
       ),
     );
+    await conn.connectionFuture;
 
     await satellite.performSnapshot();
 
@@ -1330,8 +1378,7 @@ void main() {
     final sentLsn = satellite.client.getLastSentLsn();
     expect(sentLsn, numberToBytes(1));
 
-    await satellite
-        .handleConnectivityStateChange(ConnectivityState.disconnected);
+    satellite.disconnect(null);
 
     await adapter.run(
       Statement(
@@ -1346,18 +1393,57 @@ void main() {
     expect(lsn1, sentLsn);
 
     // Once connectivity is restored, we will immediately run a snapshot to send pending rows
-    await satellite.handleConnectivityStateChange(ConnectivityState.available);
+    await satellite.connectWithBackoff();
+
     // Wait for snapshot to run
     await Future<void>.delayed(const Duration(milliseconds: 200));
     final lsn2 = satellite.client.getLastSentLsn();
     expect(lsn2, numberToBytes(2));
   });
 
+  test('notifies about JWT expiration', () async {
+    await runMigrations();
+    await startSatellite(satellite, authConfig, token);
+
+    // give some time for Satellite to start
+    // (needed because connecting and starting replication are async)
+    await Future<void>.delayed(const Duration(milliseconds: 100));
+
+    bool hasListened = false;
+    notifier.subscribeToConnectivityStateChanges((notification) {
+      if (hasListened) {
+        return;
+      }
+
+      expect(notification.dbName, dbName);
+      expect(
+        notification.connectivityState.status,
+        ConnectivityStatus.disconnected,
+      );
+      expect(
+        notification.connectivityState.reason?.code,
+        SatelliteErrorCode.authExpired,
+      );
+
+      hasListened = true;
+    });
+
+    // mock JWT expiration
+    client.emitSocketClosedError(SocketCloseReason.authExpired);
+
+    // give the notifier some time to fire
+    await Future<void>.delayed(const Duration(milliseconds: 100));
+    expect(hasListened, true);
+
+    // check that the client is disconnected
+    expect(client.isConnected(), false);
+  });
+
   test(
       'garbage collection is triggered when transaction from the same origin is replicated',
       () async {
     await runMigrations();
-    await startSatellite(satellite, authConfig);
+    await startSatellite(satellite, authConfig, token);
 
     await adapter.run(
       Statement(
@@ -1398,7 +1484,7 @@ void main() {
     final base64lsn = base64.encode(numberToBytes(kMockBehindWindowLsn));
     await satellite.setMeta('lsn', base64lsn);
     try {
-      final conn = await startSatellite(satellite, authConfig);
+      final conn = await startSatellite(satellite, authConfig, token);
       await conn.connectionFuture;
       final lsnAfter = await satellite.getMeta<String?>('lsn');
       expect(lsnAfter, isNot(base64lsn));
@@ -1417,7 +1503,7 @@ void main() {
 
     int numExpects = 0;
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await Future.wait<dynamic>(
       [
         satellite.initializing!.waitOn(),
@@ -1444,7 +1530,7 @@ void main() {
     client.setRelations(kTestRelations);
     client.setRelationData(tablename, parentRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef = ClientShapeDefinition(
@@ -1457,7 +1543,7 @@ void main() {
         await satellite.subscribe([shapeDef]);
     await synced;
 
-// first notification is 'connected'
+    // first notification is 'connected'
     expect(notifier.notifications.length, 2);
     final changeNotification = notifier.notifications[1] as ChangeNotification;
     expect(changeNotification.changes.length, 1);
@@ -1466,9 +1552,16 @@ void main() {
       Change(
         qualifiedTablename: qualified,
         rowids: [],
+        recordChanges: [
+          RecordChange(
+            primaryKey: {'id': 1},
+            type: RecordChangeType.initial,
+          ),
+        ],
       ),
     );
 
+    // wait for process to apply shape data
     try {
       final row = await adapter.query(
         Statement(
@@ -1506,7 +1599,7 @@ void main() {
     client.setRelations(kTestRelations);
     client.setRelationData(tablename, parentRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef = ClientShapeDefinition(
@@ -1536,7 +1629,7 @@ void main() {
     client.setRelations(kTestRelations);
     client.setRelationData(tablename, parentRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef = ClientShapeDefinition(
@@ -1575,7 +1668,7 @@ void main() {
     client.setRelations(kTestRelations);
     client.setRelationData(tablename, parentRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef = ClientShapeDefinition(
@@ -1628,7 +1721,7 @@ void main() {
     client.setRelations(kTestRelations);
     client.setRelationData(tablename, childRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef1 = ClientShapeDefinition(
@@ -1664,7 +1757,7 @@ void main() {
     client.setRelationData('parent', parentRecord);
     client.setRelationData(tablename, childRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef1 = ClientShapeDefinition(
@@ -1713,7 +1806,7 @@ void main() {
     client.setRelationData('parent', parentRecord);
     client.setRelationData('child', childRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final ClientShapeDefinition shapeDef1 = ClientShapeDefinition(
@@ -1766,7 +1859,7 @@ void main() {
     client.setRelationData(tablename, parentRecord);
     client.setRelationData('another', {});
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final ClientShapeDefinition shapeDef1 = ClientShapeDefinition(
@@ -1824,7 +1917,7 @@ void main() {
     client.setRelations(kTestRelations);
     client.setRelationData(tablename, parentRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef1 = ClientShapeDefinition(
@@ -1898,7 +1991,7 @@ void main() {
 
   test("Garbage collecting the subscription doesn't generate oplog entries",
       () async {
-    await startSatellite(satellite, authConfig);
+    await startSatellite(satellite, authConfig, token);
     await runMigrations();
     await adapter.run(Statement("INSERT INTO parent(id) VALUES ('1'),('2')"));
     final ts = await satellite.performSnapshot();
@@ -1931,7 +2024,7 @@ void main() {
     client.setRelations(kTestRelations);
     client.setRelationData(tablename, parentRecord);
 
-    final conn = await startSatellite(satellite, authConfig);
+    final conn = await startSatellite(satellite, authConfig, token);
     await conn.connectionFuture;
 
     final shapeDef = ClientShapeDefinition(
@@ -2033,7 +2126,7 @@ void main() {
   });
 
   test('connection backoff success', () async {
-    client.disconnect();
+    client.shutdown();
 
     int numExpects = 0;
 
@@ -2054,6 +2147,30 @@ void main() {
     );
 
     expect(numExpects, 3);
+  });
+
+  test('connection cancelled on disconnect', () async {
+    // such that satellite can't connect to Electric and will keep retrying
+    client.shutdown();
+    final conn = await startSatellite(satellite, authConfig, token);
+
+    // We expect the connection to be cancelled
+    final fut = expectLater(
+      conn.connectionFuture,
+      throwsA(
+        isA<SatelliteException>().having(
+          (e) => e.code,
+          'code',
+          SatelliteErrorCode.connectionCancelledByDisconnect,
+        ),
+      ),
+    );
+
+    // Disconnect Satellite
+    satellite.clientDisconnect();
+
+    // Await until the connection promise is rejected
+    await fut;
   });
 
   // check that performing snapshot doesn't throw without resetting the performing snapshot assertions