+++ todos_flutter/lib/auth.dart
@@ -8,7 +8,7 @@ final dummyUserId = genUUID();
 // Generate an insecure authentication JWT.
 // See https://electric-sql.com/docs/usage/auth for more details.
 String authToken() {
-  final claims = {'user_id': dummyUserId};
+  final claims = {'sub': dummyUserId};
 
   return insecureAuthToken(claims);
 }